+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ A new header is inserted each time a *tag* is created.
 
 - WebGL: reduce max instance count to work around Chrome issues [⚠️ **Recompile Materials**]
 - engine: rework material/shader sampler binding code [⚠️ **Recompile Materials**]
+- gltfio: move the API for `recomputeBoundingBoxes` [⚠️ **API Change**]
 
 ## v1.26.0
