+++ samples/gltf_viewer.cpp
@@ -89,7 +89,6 @@ struct App {
     gltfio::TextureProvider* stbDecoder = nullptr;
     gltfio::TextureProvider* ktxDecoder = nullptr;
     bool recomputeAabb = false;
-    bool ignoreBindTransform = false;
 
     bool actualSize = false;
 
@@ -139,8 +138,6 @@ static void printUsage(char* name) {
         "       Do not scale the model to fit into a unit cube\n\n"
         "   --recompute-aabb, -r\n"
         "       Ignore the min/max attributes in the glTF file\n\n"
-        "   --ignore-bind-transform, -g\n"
-        "       Ignore bind transform when recomputing aabb (use with -r)\n\n"
         "   --settings=<path to JSON file>, -t\n"
         "       Apply the settings in the given JSON file\n\n"
         "   --ubershader, -u\n"
@@ -169,7 +166,7 @@ static std::ifstream::pos_type getFileSize(const char* filename) {
 }
 
 static int handleCommandLineArguments(int argc, char* argv[], App* app) {
-    static constexpr const char* OPTSTR = "ha:i:usc:rgt:b:ev";
+    static constexpr const char* OPTSTR = "ha:i:usc:rt:b:ev";
     static const struct option OPTIONS[] = {
         { "help",         no_argument,          nullptr, 'h' },
         { "api",          required_argument,    nullptr, 'a' },
@@ -180,7 +177,6 @@ static int handleCommandLineArguments(int argc, char* argv[], App* app) {
         { "actual-size",  no_argument,          nullptr, 's' },
         { "camera",       required_argument,    nullptr, 'c' },
         { "recompute-aabb", no_argument,        nullptr, 'r' },
-        { "ignore-bind-transform", no_argument, nullptr, 'g' },
         { "settings",     required_argument,    nullptr, 't' },
         { "split-view",   no_argument,          nullptr, 'v' },
         { nullptr, 0, nullptr, 0 }
@@ -228,9 +224,6 @@ static int handleCommandLineArguments(int argc, char* argv[], App* app) {
                 break;
             case 'r':
                 app->recomputeAabb = true;
-                break;
-            case 'g':
-                app->ignoreBindTransform = true;
                 break;
             case 't':
                 app->settingsFile = arg;
@@ -429,8 +422,6 @@ int main(int argc, char** argv) {
         ResourceConfiguration configuration = {};
         configuration.engine = app.engine;
         configuration.gltfPath = gltfPath.c_str();
-        configuration.recomputeBoundingBoxes = app.recomputeAabb;
-        configuration.ignoreBindTransform = app.ignoreBindTransform;
         configuration.normalizeSkinningWeights = true;
 
         if (!app.resourceLoader) {
@@ -445,6 +436,10 @@ int main(int argc, char** argv) {
         if (!app.resourceLoader->asyncBeginLoad(app.asset)) {
             std::cerr << "Unable to start loading resources for " << filename << std::endl;
             exit(1);
+        }
+
+        if (app.recomputeAabb) {
+            app.asset->getInstance()->recomputeBoundingBoxes();
         }
 
         app.asset->releaseSourceData();
